renamed connection and schema
renamed the connection file and included the schema to be used inside the connection file instead? corrected some variables, included model which writes the user's options to database
