+++ microsvcs/gpt_generation/app/src/model/UserSchema.ts
@@ -1,11 +1,24 @@
-import { usersModel } from './mongo_conn.js';
+import mongoose from "mongoose";
+const { Schema, model, connect } = mongoose;
+import { MONGO_CREDENTIALS, MONGO_URL } from '../var/variables.js';
 
-run().catch(err => console.log(err));
+mongoose.connect(MONGO_URL, {
+  tls: true,
+  tlsCertificateKeyFile: MONGO_CREDENTIALS,
+  authMechanism: 'MONGODB-X509',
+  authSource: '$external',
+});
 
-async function run() {
-  const user = await usersModel.create({
-    uid: 'Bill',
-    scenario: 'bill@initech.com',
-  });
-  console.log(user); // 'bill@initech.com'
+interface IUserScenario {
+  uid: string;
+  scenario: string;
+  option: string;
 }
+
+const userSchema = new Schema<IUserScenario>({
+  uid: { type: String, required: true },
+  scenario: { type: String, required: true },
+  option: { type: String, required: true }
+}, { timestamps: true });
+
+export const usersModel = mongoose.model<IUserScenario>('User.Scenario', userSchema);
